CSE solutions

ConstrainedStrategicEquilibrium.CSESolution — Type

abstract type CSESolution

Base type for all CSE solutions.

ConstrainedStrategicEquilibrium.SymmetricCSESolution — Type

mutable struct SymmetricCSESolution <: ConstrainedStrategicEquilibrium.CSESolution

Contains the solution to the CSE problem.

cse::DataFrames.DataFrame: A DataFrame containing the CSE and BNE evaluated at the given points. Columns names are x, CSE(x) and BNE(x).

mse::Float64: Mean squared error of the CSE compared to the BNE

resid::Float64: Norm of the derivatives

knot::DataFrames.DataFrame

alph_bet::DataFrames.DataFrame

x_f::DataFrames.DataFrame

success::Bool: Whether the calculation was successful or not

c_1::Float64: Stop criteria C_1 (comparison of the CSE at the current value of n with the CSE at the previous value of n)

c_2::Float64: Stop criteria C_2 (norm of the residual)

problem::ConstrainedStrategicEquilibrium.SymmetricCSEProblem: The problem that this solution was generated for

n::Int64: The value of n used in this solution

u::Array{Float64}: The data used in generating the solution

solver_solution::Union{Nothing, SciMLBase.NonlinearSolution}: The solution object returned by the solver

ConstrainedStrategicEquilibrium.AsymmetricCSESolution — Type

mutable struct AsymmetricCSESolution <: ConstrainedStrategicEquilibrium.CSESolution

Contains the solution to the asymmetric CSE problem.

cse::DataFrames.DataFrame: A DataFrame containing the CSE and BNE evaluated at the given points

resid::Float64: Norm of the derivatives

knot::Dict{Symbol, DataFrames.DataFrame}: The knots for both bidders

alph_bet::Dict{Symbol, DataFrames.DataFrame}: Alpha and beta values for the piecewise linear functions for both bidders

x_f::Dict{Symbol, DataFrames.DataFrame}: x values and derivatives at the final point

norm_derivatives::Dict{Symbol, Float64}: Norm of the derivatives for each player

success::Bool: Whether the calculation was successful or not

nfeval::Int64: Number of function evaluations

njacs::Int64: Number of Jacobians created during the solve

nfactors::Int64: Number of factorisations of the Jacobian required for the solve

nsolve::Int64: Number of linear solves required for the solve

nsteps::Int64: Total number of iterations for the nonlinear solver

c_1::Dict{Symbol, Float64}: Stop criteria C_1 (comparison of CSE with previous n value) for each player

c_2::Float64: Stop criteria C_2 (norm of the residual)

problem::ConstrainedStrategicEquilibrium.AsymmetricCSEProblem: The problem that this solution was generated for

n::Int64: The value of n used in this solution

u::Array{Float64}: The data used in generating the solution

solver_solution::Union{Nothing, SciMLBase.NonlinearSolution}: The solution object returned by the solver